test code

+++ primary/primary.h
@@ -16,7 +16,7 @@ void i2s(int i,char *s);
 int init_sdcard(void);
 int sdcard_open(uint8_t hours, uint8_t minutes, uint8_t seconds);
 void sdcard_close();
-void get_steps(uint8_t *points, int size, uint8_t * avg, uint8_t * steps, uint8_t * activity_level);
+void get_steps(uint8_t *points, uint8_t size, uint8_t * avg, uint8_t * steps, uint8_t * activity_level);
 
 //test
 #endif
